my 1st commite

+++ src/LogicalPrograme/FactorialNumberCheck.java
@@ -6,7 +6,7 @@ public class FactorialNumberCheck {
         //!8= 8*7*6*5*4*3*2*1 = 40320
         //!4= 4*3*2*1 = 24
         int no = 8;
-        int fact = 1;
+        long fact = 1;
         for (int i = 1; i <= no; i++) {
             fact = fact * i;
         }
